Refactor scraping classes
