+++ lib/scraping/scraping_exception.rb
